requirements.txt

# First level requirements
pillow>=5.4.1
